+++ build.gradle
@@ -38,13 +38,14 @@ repositories {
 }
 
 dependencies {
-    compile 'org.testng:testng:6.13'
+    compile 'org.testng:testng:6.14.3'
     compile 'com.google.guava:guava:23.5-jre'
     compile 'org.apache.commons:commons-lang3:3.7'
     compile 'com.github.wmacevoy:kiss:1.0'
     compile 'org.projectlombok:lombok:1.16.20'
     compile 'one.util:streamex:0.6.6'
     compile 'io.vavr:vavr:0.9.2'
+    compile 'com.google.inject:guice:4.1.0:no_aop'
 
     testCompile 'org.mockito:mockito-core:2.17.0'
     testCompile 'org.assertj:assertj-core:3.9.1'
@@ -65,7 +66,8 @@ test{
     useTestNG()
 
     filter{
-        includeTestsMatching 'TestSuite'
+        includeTestsMatching 'unit.*'
+        includeTestsMatching 'end.to.end.*'
     }
 }
