+++ src/test/java/util/InvokedMethodNameListener.java
@@ -1,3 +1,5 @@
+package util;
+
 import one.util.streamex.StreamEx;
 import org.testng.*;
