+++ src/test/java/util/TestUtils.java
@@ -1,3 +1,5 @@
+package util;
+
 import lombok.experimental.UtilityClass;
 import org.testng.ITestNGListener;
 import org.testng.TestNG;
